Remove Pip from default includes
This commit disables pip installation from windows builds.  pip can be
installed at any time by running `python -mensurepip` this will install
pip and setuptools wheels that are packaged with the core.

The decision to remove pip from the core was due to our additive model
of packages - a pip or setuptools built in the build process is unable
to entirely replace the original modules.