Fix Async import problem on Posix
Windows only seems to return WSAE* errors, so we need to check for these
errors on Windows.

However the WSAE* error definitions only exist on Windows, they do not
exist on Posix.

Trying to import them causes a failure on Posix.

On Posix, we now just reassign the WSAE* error values to the normal E*
values, letting the code compile.