Move vcruntime to main folder
Remove vcruntime from the main folder, and remove vcruntime from the
DLLs subfolder. I tried using the existing apparatus, but matching a
substring from the exlusion list to the src string does not work as
expected, possibly because of the difference in directories, so I
have flipped the substring matching to look for vcruntime within the
src. It doesn't look as nice, but it produces the correct result.